+++ online_shop/api/admin.py
@@ -1,3 +1,12 @@
 from django.contrib import admin
+from .models import Address, Customer, Category, Shop, Product, ProductColor, ProductImage,  Order
 
-# Register your models here.
+
+admin.site.register(Address)
+admin.site.register(Customer)
+admin.site.register(Category)
+admin.site.register(Shop)
+admin.site.register(Product)
+admin.site.register(ProductColor)
+admin.site.register(ProductImage)
+admin.site.register(Order)
