+++ online_shop/online_shop/urls.py
@@ -15,10 +15,10 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
-
+from django.conf.urls.static import static
+from .settings import MEDIA_URL, MEDIA_ROOT
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('shop.urls')),
     path('api/', include('api.urls')),
-]
+] + static(MEDIA_URL, document_root=MEDIA_ROOT)
