--- online_shop/shop/admin.py
@@ -1,12 +1,0 @@
-from django.contrib import admin
-from .models import Address, Customer, Category, Shop, Product, ProductColor, ProductImage,  Order
-
-
-admin.site.register(Address)
-admin.site.register(Customer)
-admin.site.register(Category)
-admin.site.register(Shop)
-admin.site.register(Product)
-admin.site.register(ProductColor)
-admin.site.register(ProductImage)
-admin.site.register(Order)
