--- online_shop/shop/templates/registration.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Title</title>
-</head>
-<body>
-    <div class="row" style="margin-bottom: 50px; margin-top: 50px">
-        <div class="col-md-6 offset-md-3">
-            <h3 class='text-center'>Регистрация нового пользователя</h3>
-            <hr>
-            <form action="" method="POST">
-                {% csrf_token %}
-                {% for field in form %}
-                    <div class="form-group {% if field.errors %}alert alert-danger{% endif %}">
-                        {{ field.errors }}
-                        {{ field.label_tag }}
-                        {{ field }}
-
-                        {% if field.help_text %}
-                        <small class="text-muted">
-                            {{ field.help_text }}
-                        </small>
-
-                {% endif %}
-                    </div>
-                {% endfor %}
-                <input type="submit" class="btn btn-success" value="Войти">
-            </form>
-        </div>
-    </div>
-</body>
-</html>
