--- online_shop/shop/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-from .views import RegistrationView
-
-urlpatterns = [
-    # Auth
-    path('signup/', RegistrationView.as_view(), name='signup')
-]
